sender/apt-requirements.txt
libpam0g-dev
python3-venv
python3-pip
python3-evdev
avahi-utils